requirements.txt
ipykernel==6.20.2
ipython==8.8.0
LinearRegression==0.0.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
numpy==1.24.1
Pillow==9.4.0
scikit-learn==1.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.0
torch==2.0.1
torchvision==0.15.2
